moved scheduled task to a separate class MentionService only deals with processing mentions & saving
